+++ src/sphinx_substitution_extensions/__init__.py
@@ -5,13 +5,20 @@ Custom Sphinx extensions.
 from typing import Any, ClassVar
 
 from beartype import beartype
-from docutils.nodes import Element, Node, system_message
+from docutils.nodes import (
+    Element,
+    Node,
+    substitution_definition,
+    system_message,
+)
 from docutils.parsers.rst import directives
 from docutils.parsers.rst.roles import code_role
 from docutils.parsers.rst.states import Inliner
 from docutils.statemachine import StringList
+from myst_parser.mocking import MockInliner
 from sphinx import addnodes
 from sphinx.application import Sphinx
+from sphinx.config import Config
 from sphinx.directives.code import CodeBlock
 from sphinx.environment import BuildEnvironment
 from sphinx.roles import XRefRole
@@ -20,6 +27,61 @@ from sphinx.util.typing import ExtensionMetadata, OptionSpec
 from sphinx_substitution_extensions.shared import (
     SUBSTITUTION_OPTION_NAME,
 )
+
+
+def _get_delimiter_pairs(
+    env: BuildEnvironment,
+    config: Config,
+) -> set[tuple[str, str]]:
+    """
+    Get the delimiter pairs for substitution.
+    """
+    markdown_suffixes = {
+        key.lstrip(".")
+        for key, value in config.source_suffix.items()
+        if value == "markdown"
+    }
+
+    # Use `| |` on reST as it is the default substitution syntax.
+    # Use `| |` on MyST for backwards compatibility as this is what we
+    # originally shipped with.
+    delimiter_pairs = {("|", "|")}
+    parser_supported_formats = set(env.parser.supported)
+    if parser_supported_formats.intersection(markdown_suffixes):
+        opening_delimiter, closing_delimiter = config.myst_sub_delimiters
+        new_delimiter_pair = (
+            opening_delimiter + opening_delimiter,
+            closing_delimiter + closing_delimiter,
+        )
+        delimiter_pairs = {*delimiter_pairs, new_delimiter_pair}
+
+    return delimiter_pairs
+
+
+def _get_substitution_defs(
+    env: BuildEnvironment,
+    config: Config,
+    substitution_defs: dict[str, substitution_definition],
+) -> dict[str, str]:
+    """
+    Get the substitution definitions from the environment.
+    """
+    markdown_suffixes = {
+        key.lstrip(".")
+        for key, value in config.source_suffix.items()
+        if value == "markdown"
+    }
+
+    parser_supported_formats = set(env.parser.supported)
+    if parser_supported_formats.intersection(markdown_suffixes):
+        if "substitution" in config.myst_enable_extensions:
+            return dict(config.myst_substitutions)
+    else:
+        return {
+            key: value.astext() for key, value in substitution_defs.items()
+        }
+
+    return {}
 
 
 @beartype
@@ -37,35 +99,16 @@ class SubstitutionCodeBlock(CodeBlock):
         """
         new_content = StringList()
         existing_content = self.content
-        substitution_defs = {}
+        substitution_defs = _get_substitution_defs(
+            env=self.env,
+            config=self.config,
+            substitution_defs=self.state.document.substitution_defs,
+        )
 
-        markdown_suffixes = {
-            key.lstrip(".")
-            for key, value in self.config.source_suffix.items()
-            if value == "markdown"
-        }
-
-        # Use `| |` on reST as it is the default substitution syntax.
-        # Use `| |` on MyST for backwards compatibility as this is what we
-        # originally shipped with.
-        delimiter_pairs = {("|", "|")}
-        parser_supported_formats = set(self.env.parser.supported)
-        if parser_supported_formats.intersection(markdown_suffixes):
-            if "substitution" in self.config.myst_enable_extensions:
-                substitution_defs = self.config.myst_substitutions
-            opening_delimiter, closing_delimiter = (
-                self.config.myst_sub_delimiters
-            )
-            new_delimiter_pair = (
-                opening_delimiter + opening_delimiter,
-                closing_delimiter + closing_delimiter,
-            )
-            delimiter_pairs = {*delimiter_pairs, new_delimiter_pair}
-        else:
-            substitution_defs = {
-                key: value.astext()
-                for key, value in self.state.document.substitution_defs.items()
-            }
+        delimiter_pairs = _get_delimiter_pairs(
+            env=self.env,
+            config=self.config,
+        )
 
         for item in existing_content:
             new_item = item
@@ -101,7 +144,7 @@ class SubstitutionCodeRole:
         rawtext: str,
         text: str,
         lineno: int,
-        inliner: Inliner,
+        inliner: Inliner | MockInliner,
         # We allow mutable defaults as the Sphinx implementation requires it.
         options: dict[Any, Any] = {},  # noqa: B006
         content: list[str] = [],  # noqa: B006
@@ -109,12 +152,43 @@ class SubstitutionCodeRole:
         """
         Replace placeholders with given variables.
         """
-        inliner_document = inliner.document
-        for name, value in inliner_document.substitution_defs.items():
-            replacement = value.astext()
-            text = text.replace(f"|{name}|", replacement)
-            rawtext = text.replace(f"|{name}|", replacement)
-            rawtext = rawtext.replace(name, replacement)
+        settings = inliner.document.settings
+        env = settings.env  # pyright: ignore[reportUnknownMemberType, reportUnknownVariableType]
+        assert isinstance(env, BuildEnvironment)
+        substitution_defs = _get_substitution_defs(
+            env=env,
+            config=env.config,
+            substitution_defs=inliner.document.substitution_defs,
+        )
+
+        delimiter_pairs = _get_delimiter_pairs(
+            env=env,
+            config=env.config,
+        )
+
+        for name, value in substitution_defs.items():
+            for delimiter_pair in delimiter_pairs:
+                opening_delimiter, closing_delimiter = delimiter_pair
+                text = text.replace(
+                    f"{opening_delimiter}{name}{closing_delimiter}",
+                    value,
+                )
+                rawtext = text.replace(
+                    f"{opening_delimiter}{name}{closing_delimiter}",
+                    value,
+                )
+                rawtext = rawtext.replace(name, value)
+
+        # ``types-docutils`` says that ``code_role`` requires an ``Inliner``
+        # for ``inliner``.
+        #
+        # We can remove this when
+        # https://github.com/executablebooks/MyST-Parser/issues/1017
+        # is resolved by typing ``inliner`` as ``Inliner``.
+        if isinstance(inliner, MockInliner):
+            new_inliner = Inliner()
+            new_inliner.document = inliner.document
+            inliner = new_inliner
 
         return code_role(
             role=typ,